requirements.txt
flask
flask-cors
flask-mail
python-dotenv
pymongo
google-search-results
joblib
numpy
pandas
scikit-learn==1.6.1
pickle-mixin==1.0.2
setuptools
wheel
gdown
pymongo==4.6.3
dnspython==2.4.2
pymongo[srv]
huggingface_hub
